fix: remove pinning of zfs compatible kernel
This is not supported by upstream anymore, and we are expected to just
default to the stable kernel.

evaluation warning: zfs.latestCompatibleLinuxPackages is deprecated and
is now pointing at the default kernel. If using the stable LTS kernel
(default `linuxPackages` is not possible then you must explicitly pin a
specific kernel release. For example,
`boot.kernelPackages = pkgs.linuxPackages_6_6`. Please be aware that
non-LTS kernels are likely to go EOL before ZFS supports the latest
supported non-LTS release, requiring manual intervention.