firefly-iii-auto-importer: change cron schedule to 07:00
For some reason I cannot understand, banks like to go offline at
midnight...
Hopefully they are available in the morning.

This was set to a few minutes before midnight so that the importer could
only get that day's transactions, but the importer can now reliably
filter out duplicates, so we are getting the transactions from last week
anyway.