fix(nextcloud): error after activation
Due to PHP's realpath cache, every time the activation scripts run,
Nextcloud stops working for a brief moment. This happens because the
secrets handled by agenix change their path, and PHP does not follow the
new destination of the symlink until the cache expires.

This commit adds an activation script that reloads php-fpm after agenix
has updated secrets, so that it clears the cache.